backend/pyproject.toml
[tool.poetry]
name = "clinical-health-platform"
version = "0.1.0"
description = "Backend for Clinical Healthcare Platform focusing on depression treatment with executive control network dysfunction"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
pydantic = "^2.4.2"
alembic = "^1.12.0"
python-dotenv = "^1.0.0"
aiosqlite = "^0.19.0"
email-validator = "^2.1.0"
greenlet = "^3.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.0"
black = "^23.9.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
